# resources/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Continy CMS API",
    "description": "RESTful API for Continy - A headless CMS with MongoDB backend",
    "version": "1.0.0",
    "contact": {
      "name": "Continy Support"
    }
  },
  "servers": [],
...
  "paths": {
...
    "/collections/{collection}": {
      "get": {
        "tags": ["Collections"],
        "summary": "Get collection",
        "description": "Get a specific collection's details with contents",
        "operationId": "getCollection",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/CollectionId"}
        ],
        "responses": {
          "200": {
            "description": "Collection details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionDataResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Collections"],
        "summary": "Update collection",
        "description": "Update a collection's information",
        "operationId": "updateCollection",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/CollectionId"}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collection updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Collections"],
        "summary": "Delete collection",
        "description": "Delete a collection (must have no contents)",
        "operationId": "deleteCollection",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/CollectionId"}
        ],
        "responses": {
          "200": {
            "description": "Collection deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Collection has contents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/contents/{content}": {
      "get": {
        "tags": ["Contents"],
        "summary": "Get content",
        "description": "Get a specific content's details with collection and recent versions",
        "operationId": "getContent",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/ContentId"}
        ],
        "responses": {
          "200": {
            "description": "Content details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDataResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Contents"],
        "summary": "Update content",
        "description": "Update a content's information",
        "operationId": "updateContent",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/ContentId"}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Content updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Contents"],
        "summary": "Delete content",
        "description": "Delete a content",
        "operationId": "deleteContent",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/ContentId"}
        ],
        "responses": {
          "200": {
            "description": "Content deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Bearer token obtained from the login endpoint"
      }
    },
    "parameters": {
...
      "CollectionId": {
        "name": "collection",
        "in": "path",
        "required": true,
        "description": "Collection slug (URL-friendly identifier, e.g. 'blog-posts')",
        "schema": {"type": "string"},
        "example": "blog-posts"
      },
      "ContentId": {
        "name": "content",
        "in": "path",
        "required": true,
        "description": "Content ID (MongoDB ObjectId)",
        "schema": {"type": "string"}
      },
...
    },
    "schemas": {
...
      "Collection": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "description": "MongoDB ObjectId (internal identifier)"},
          "name": {"type": "string", "description": "Display name of the collection"},
          "slug": {"type": "string", "description": "URL-friendly identifier used in API routes (e.g. 'blog-posts')"},
          "description": {"type": "string", "nullable": true},
          "schema": {"type": "object", "description": "Collection schema definition"},
          "settings": {"type": "object", "description": "Collection settings"},
          "collection_meta": {"type": "object", "nullable": true},
          "contents_count": {"type": "integer"},
          "contents": {"type": "array", "items": {"$ref": "#/components/schemas/Content"}},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "Content": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "description": "MongoDB ObjectId (used in API routes)"},
          "collection_id": {"type": "string", "description": "MongoDB ObjectId of the parent collection"},
          "title": {"type": "string"},
          "slug": {"type": "string"},
          "status": {"type": "string", "enum": ["draft", "published", "archived"]},
          "status_label": {"type": "string"},
          "is_published": {"type": "boolean"},
          "current_version": {"type": "integer"},
          "published_version_id": {"type": "string", "nullable": true},
          "elements": {"type": "array", "items": {"$ref": "#/components/schemas/Element"}},
          "metadata": {"type": "object"},
          "collection": {"$ref": "#/components/schemas/Collection"},
          "versions": {"type": "array", "items": {"$ref": "#/components/schemas/ContentVersion"}},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "Element": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "type": {"type": "string", "enum": ["text", "heading", "image", "video", "quote", "code", "list", "table", "embed", "wrapper"]},
          "data": {"type": "object"},
          "order": {"type": "integer"},
          "parent_id": {"type": "string", "nullable": true}
        }
      },
      "ContentVersion": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "version": {"type": "integer"},
          "title": {"type": "string"},
          "elements": {"type": "array", "items": {"$ref": "#/components/schemas/Element"}},
          "metadata": {"type": "object"},
          "change_note": {"type": "string", "nullable": true},
          "created_by": {"type": "integer"},
          "created_at": {"type": "string", "format": "date-time"}
        }
      },
...
      "UpdateCollectionRequest": {
        "type": "object",
        "properties": {
          "name": {"type": "string"},
          "slug": {"type": "string"},
          "description": {"type": "string"},
          "schema": {"type": "object"},
          "settings": {"type": "object"}
        }
      },
...
      "UpdateContentRequest": {
        "type": "object",
        "properties": {
          "title": {"type": "string"},
          "slug": {"type": "string"},
          "elements": {"type": "array", "items": {"$ref": "#/components/schemas/Element"}},
          "metadata": {"type": "object"},
          "change_note": {"type": "string"}
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {"type": "string"}
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {"type": "string"}
        }
      },
...
      "CollectionResponse": {
        "type": "object",
        "properties": {
          "message": {"type": "string"},
          "data": {"$ref": "#/components/schemas/Collection"}
        }
      },
      "CollectionDataResponse": {
        "type": "object",
        "properties": {
          "data": {"$ref": "#/components/schemas/Collection"}
        }
      },
...
      "ContentResponse": {
        "type": "object",
        "properties": {
          "message": {"type": "string"},
          "data": {"$ref": "#/components/schemas/Content"}
        }
      },
      "ContentDataResponse": {
        "type": "object",
        "properties": {
          "data": {"$ref": "#/components/schemas/Content"}
        }
      },
...
    }
  }
}